spbackend/requirements.txt
asgiref==3.8.1
Django==4.2
django-debug-toolbar==4.3.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
importlib_metadata==8.4.0
Markdown==3.7
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.20.1
